Move alert endpoints to /api/v1 + first version of podman compose with nginx
