+++ config/oauth2-proxy.cfg.example
@@ -107,9 +107,6 @@ skip_auth_routes = [
 ## mark paths as API routes to get HTTP Status code 401 instead of redirect to login page
 api_routes = [
   "^/api",
-  "^/alerts",
-  "^/alert-types", 
-  "^/session-id",
   "^/ws"
 ]
