+++ podman-compose.yml
@@ -23,12 +23,11 @@ services:
     volumes:
       - ./config/oauth2-proxy-container.cfg:/config/oauth2-proxy-container.cfg:ro
       - ./config/templates:/templates:ro  # Mount custom templates
-    ports:
-      - "4180:4180"  # Direct OAuth2-proxy access
+    # No direct port exposure - accessed only through reverse proxy
     depends_on:
-      - tarsy-backend
+      - backend
 
-  tarsy-backend:
+  backend:
     build:
       context: ./backend
       dockerfile: Dockerfile
@@ -46,26 +45,27 @@ services:
       database:
         condition: service_healthy
     
-  tarsy-frontend:
+  dashboard:
     build:
       context: ./dashboard
       dockerfile: Dockerfile
-    command: ["npx", "vite", "dev", "--host", "0.0.0.0", "--port", "3000"]
-    ports:
-      - "3000:3000"  # Direct frontend access
-    environment:
-      - NODE_ENV=development
-      # Development server configuration (what browser uses to access frontend)
-      - VITE_DEV_SERVER_HOST=localhost  # Browser access URL
-      - VITE_DEV_SERVER_PORT=3000
-      # Production URLs (only used in production builds, ignored in dev mode)
-      - VITE_API_BASE_URL=http://localhost:4180
-      - VITE_WS_BASE_URL=ws://localhost:4180
-      # Vite proxy targets (container-to-container communication)
-      - VITE_PROXY_TARGET_HTTP=http://oauth2-proxy:4180  # Internal proxy target
-      - VITE_PROXY_TARGET_WS=ws://oauth2-proxy:4180      # Internal proxy target  
-      - VITE_PROXY_HOST_HEADER=localhost:4180            # For OAuth2 cookie compatibility
+      args:
+        # Build-time environment variables for production build
+        - VITE_API_BASE_URL=http://localhost:80
+        - VITE_WS_BASE_URL=ws://localhost:80
+    # No direct ports exposed - accessed only through reverse proxy
+    # nginx serves React production build on port 80 inside container
     depends_on:
+      - oauth2-proxy
+
+  reverse-proxy:
+    image: mirror.gcr.io/library/nginx:alpine
+    ports:
+      - "80:80"  # Main access point
+    volumes:
+      - ./config/nginx-reverse-proxy.conf:/etc/nginx/conf.d/default.conf:ro
+    depends_on:
+      - dashboard
       - oauth2-proxy
 
 volumes:
